Created Index File and JS File
I have created the index file as well as the javascipt file. Index file links the JS file in the body.
In the JS File, 2 function have been defined, one is for computer choice and the other is
for human choice. Computer choice uses the random function to return the computers random
choice. While the human choice prompts a user for their choice.